requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
Brlapi==0.6.6
businesstime==0.3.0
certifi==2018.11.29
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
Deprecated==1.2.4
distro-info==0.18
httplib2==0.9.2
idna==2.8
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
meld==3.18.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pdfkit==0.6.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGithub==1.43.5
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pystache==0.5.4
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.2.0
wadllib==1.3.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
